HW5
change casting in ListStorage
remove abstract method from AbstractArrayStorage

+++ src/ru/webapp/basejava/storage/ListStorage.java
@@ -40,7 +40,7 @@ public class ListStorage extends AbstractStorage {
 
     @Override
     protected void doDelete(Object searchKey) {
-        storage.remove((int) searchKey);
+        storage.remove(((Integer) searchKey).intValue());
     }
 
     @Override
